[KFI] Refactor.
